+++ lib/src/radial_gauge/data/data.dart
@@ -4,5 +4,4 @@ export 'gauge_axis_transformer.dart';
 export 'gauge_label_provider.dart';
 export 'gauge_border.dart';
 export 'gauge_zone.dart';
-export 'gauge_segment.dart';
 export 'radial_gauge_layout.dart';