+++ lib/src/radial_gauge/data/gauge_segment.dart
@@ -1,5 +1,0 @@
-import 'gauge_zone.dart';
-
-/// Renamed to [GaugeZone].
-@Deprecated('Renamed to GaugeZone. Will be removed in 0.6.0.')
-typedef GaugeSegment = GaugeZone;